+++ packages/blade/src/components/Card/Card.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import type { GestureResponderEvent } from 'react-native';
 import { CardSurface } from './CardSurface';
 import { CardProvider, useVerifyInsideCard } from './CardContext';
 import { LinkOverlay } from './LinkOverlay';
@@ -15,6 +16,7 @@ import type { Elevation } from '~tokens/global';
 import type { BoxProps } from '~components/Box';
 import { makeAccessible } from '~utils/makeAccessible';
 import { useVerifyAllowedChildren } from '~utils/useVerifyAllowedChildren/useVerifyAllowedChildren';
+import type { Platform } from '~utils';
 import { isReactNative } from '~utils';
 import type { Theme } from '~components/BladeProvider';
 import type { DotNotationToken } from '~utils/lodashButBetter/get';
@@ -85,11 +87,20 @@ export type CardProps = {
    */
   height?: BoxProps['height'];
   /**
+   * Sets minimum height of the card
+   */
+  minHeight?: BoxProps['minHeight'];
+  /**
+   * Sets minimum width of the card
+   */
+  minWidth?: BoxProps['minWidth'];
+  /**
    * If `true`, the card will be in selected state
    * Card will have a primary color border around it.
    *
    * @default false
    */
+
   isSelected?: boolean;
   /**
    * Makes the Card linkable by setting the `href` prop
@@ -126,7 +137,12 @@ export type CardProps = {
   /**
    * Callback triggered when the card is clicked
    */
-  onClick?: () => void;
+  onClick?: (
+    event: Platform.Select<{
+      web: React.MouseEvent;
+      native: GestureResponderEvent;
+    }>,
+  ) => void;
   /**
    * Sets the HTML element for the Card
    *
@@ -148,6 +164,8 @@ const Card = ({
   padding = 'spacing.7',
   width,
   height,
+  minHeight,
+  minWidth,
   onClick,
   isSelected = false,
   accessibilityLabel,
@@ -193,6 +211,8 @@ const Card = ({
         onClick={isReactNative() ? onClick : undefined}
         width={width}
         height={height}
+        minHeight={minHeight}
+        minWidth={minWidth}
         href={href}
         accessibilityLabel={accessibilityLabel}
         {...metaAttribute({ name: MetaConstants.Card, testID })}
@@ -200,6 +220,7 @@ const Card = ({
       >
         <CardSurface
           height={height}
+          minHeight={minHeight}
           padding={padding}
           borderRadius={borderRadius}
           elevation={elevation}